+++ inc/utils/helpers.php
@@ -12,21 +12,19 @@ class Helpers {
 	/**
 	 * Get paged value
 	 */
-	// if ( ! function_exists('get_page_num')) :
+	function get_page_num() {
 
-		function get_page_num() {
-
-			if ( get_query_var('paged') ) {
-				$paged = intval(get_query_var('paged'));
-			} else if ( get_query_var('page') ) {
-				$paged = intval(get_query_var('page'));
-			} else {
-				$paged = 1;
-			}
-
-			return $paged;
+		if ( get_query_var('paged') ) {
+			$paged = intval(get_query_var('paged'));
+		} else if ( get_query_var('page') ) {
+			$paged = intval(get_query_var('page'));
+		} else {
+			$paged = 1;
 		}
-	// endif;
+
+		return $paged;
+	}
+
 
 
 
@@ -34,54 +32,51 @@ class Helpers {
 	/**
 	 * Get value from array key
 	 */
-	// if ( ! function_exists('get_key')) :
+	static function get_key($k, $a = false) {
 
-		static function get_key($k, $a = false) {
-
-			global $widget, $cs;
-			$val = false;
+		global $widget, $cs;
+		$val = false;
 
 
-			/**
-			 * Falback for old way of getting things
-			 */
-			if ( is_array($k) || is_object($k) ) :
+		/**
+		 * Falback for old way of getting things
+		 */
+		if ( is_array($k) || is_object($k) ) :
 
-				$key   = $a;
-				$array = $k;
-			else:
+			$key   = $a;
+			$array = $k;
+		else:
 
-				$key   = $k;
-				$array = $a;
-			endif;
-
-
-			/**
-			 * Get value from froots/widget variable
-			 */
-			if ( ! $array ) :
-
-				if ( is_array($widget) ) {
-					$array = $widget;
-				}
-				elseif ( is_array($cs)) {
-					$array = $cs;
-				}
-			endif;
+			$key   = $k;
+			$array = $a;
+		endif;
 
 
-			if (!is_array($array) && !is_object($array))
-				return false;
+		/**
+		 * Get value from froots/widget variable
+		 */
+		if ( ! $array ) :
 
-			if (is_array($array) && array_key_exists($key, $array))
-				$val = $array[$key];
+			if ( is_array($widget) ) {
+				$array = $widget;
+			}
+			elseif ( is_array($cs)) {
+				$array = $cs;
+			}
+		endif;
 
-			if (is_object($array) && property_exists($array, $key))
-				$val = $array->$key;
 
-			return is_string($val) ? trim($val) : $val;
-		}
-	// endif;
+		if (!is_array($array) && !is_object($array))
+			return false;
+
+		if (is_array($array) && array_key_exists($key, $array))
+			$val = $array[$key];
+
+		if (is_object($array) && property_exists($array, $key))
+			$val = $array->$key;
+
+		return is_string($val) ? trim($val) : $val;
+	}
 
 
 
@@ -89,39 +84,36 @@ class Helpers {
 	/**
 	 * Get value from array/objec property and echo along with tag and class
 	 */
-	// if ( ! function_exists('the_key') ) :
+	function the_key( $key, $tag = false, $class = false, $array = false ) {
 
-		function the_key( $key, $tag = false, $class = false, $array = false ) {
-
-			if ( is_array($array) || is_object($array) ) {
-				$val = get_key($key, $array);
-			}
-			else {
-				$val = get_key($key);
-			}
-
-			if ( ! $val )
-				return;
-
-
-			if ( ! $tag ) {
-
-				echo $val;
-				return;
-			}
-
-			$open = '<' . $tag;
-
-			if ( $class ) {
-				$open .= ' class="'. $class .'"';
-			}
-
-			$open .= '>' . $val;
-			$open .=  '</' . $tag . '>';
-
-			echo $open;
+		if ( is_array($array) || is_object($array) ) {
+			$val = get_key($key, $array);
 		}
-	// endif;
+		else {
+			$val = get_key($key);
+		}
+
+		if ( ! $val )
+			return;
+
+
+		if ( ! $tag ) {
+
+			echo $val;
+			return;
+		}
+
+		$open = '<' . $tag;
+
+		if ( $class ) {
+			$open .= ' class="'. $class .'"';
+		}
+
+		$open .= '>' . $val;
+		$open .=  '</' . $tag . '>';
+
+		echo $open;
+	}
 
 
 
@@ -129,12 +121,9 @@ class Helpers {
 	/**
 	 * Strip all but numbers
 	 */
-	// if ( ! function_exists('get_tel')) :
-
-		function get_tel( $tel ) {
-			return preg_replace("/[^0-9]/","",$tel);
-		}
-	// endif;
+	function get_tel( $tel ) {
+		return preg_replace("/[^0-9]/","",$tel);
+	}
 
 
 
@@ -142,13 +131,10 @@ class Helpers {
 	/**
 	 * Trim text to X words
 	 */
-	// if ( ! function_exists('cs_trim_word')) :
-
-		function cs_trim_word( $text, $length ) {
-			$trimmed = wp_trim_words( $text, $num_words = $length, $more = null );
-			return $trimmed;
-		}
-	// endif;
+	function cs_trim_word( $text, $length ) {
+		$trimmed = wp_trim_words( $text, $num_words = $length, $more = null );
+		return $trimmed;
+	}
 
 
 
@@ -156,27 +142,24 @@ class Helpers {
 	/**
 	 * Trim text to X chars and append elipsis symbol
 	 */
-	// if ( ! function_exists('cs_trim_chars')) :
+	function cs_trim_chars( $text, $length = 45, $append = '&hellip;' ) {
 
-		function cs_trim_chars( $text, $length = 45, $append = '&hellip;' ) {
+		$length = (int) $length;
+		$text   = trim( strip_tags( $text ) );
 
-			$length = (int) $length;
-			$text   = trim( strip_tags( $text ) );
+		if ( strlen( $text ) > $length ) {
+			$text  = substr( $text, 0, $length + 1 );
+			$words = preg_split( "/[\s]|&nbsp;/", $text, -1, PREG_SPLIT_NO_EMPTY );
+			preg_match( "/[\s]|&nbsp;/", $text, $lastchar, 0, $length );
 
-			if ( strlen( $text ) > $length ) {
-				$text  = substr( $text, 0, $length + 1 );
-				$words = preg_split( "/[\s]|&nbsp;/", $text, -1, PREG_SPLIT_NO_EMPTY );
-				preg_match( "/[\s]|&nbsp;/", $text, $lastchar, 0, $length );
+			if ( empty( $lastchar ) )
+				array_pop( $words );
 
-				if ( empty( $lastchar ) )
-					array_pop( $words );
-
-				$text = implode( ' ', $words ) . $append;
-			}
-
-			return $text;
+			$text = implode( ' ', $words ) . $append;
 		}
-	// endif;
+
+		return $text;
+	}
 
 
 
@@ -184,29 +167,25 @@ class Helpers {
 	/**
 	 * Debug mail
 	 */
-	// if ( ! function_exists('debug_wpmail')) :
+	function debug_wpmail( $re ) {
+		if ( ! $re ) {
 
-		function debug_wpmail( $re ) {
+			global $ts_mail_errors;
+			global $phpmailer;
 
-			if ( ! $re ) {
+			if ( ! isset($ts_mail_errors) )
+				$ts_mail_errors = array();
 
-				global $ts_mail_errors;
-				global $phpmailer;
-
-				if ( ! isset($ts_mail_errors) )
-					$ts_mail_errors = array();
-
-				if ( isset($phpmailer) ) {
-					$ts_mail_errors[] = $phpmailer->ErrorInfo;
-				}
-
-				print_r('<pre>');
-				print_r($ts_mail_errors);
-				print_r('</pre>');
+			if ( isset($phpmailer) ) {
+				$ts_mail_errors[] = $phpmailer->ErrorInfo;
 			}
-		}
 
-	// endif;
+			print_r('<pre>');
+			print_r($ts_mail_errors);
+			print_r('</pre>');
+		}
+	}
+
 
 
 
@@ -255,83 +234,48 @@ class Helpers {
 
 
 	/**
-	 * Debug
-	 * Print_r with <pre> tags for readable output
-	 */
-	// if ( ! function_exists('printaj')) :
-
-		function printaj( $var, $return = false ) {
-			print_r('<pre>');
-			print_r($var, $return);
-			print_r('</pre>');
-		}
-
-	// endif;
-
-
-
-
-	/**
-	 * Debug
-	 * var_dump with <pre> tags for readable output
-	 */
-	// if ( ! function_exists('dumpaj')) :
-
-		function dumpaj( $var, $return = false ) {
-			var_dump('<pre>');
-			var_dump($var, $return);
-			var_dump('</pre>');
-		}
-	// endif;
-
-
-
-
-	/**
 	 * Easy include SVG (or PNG) icons by icon filename in assets folder
 	 */
-	// if ( ! function_exists('svg_icon')) :
-		function svg_icon( $icon = false, $return = false, $original = false ) {
+	function svg_icon( $icon = false, $return = false, $original = false ) {
 
-			if ( ! $icon ) {
-				$icon = get_key('select_icon');
-			}
-
-			if ( is_array($icon) ) {
-				$icon = $icon['select_icon'];
-			}
-
-			if ( strpos($icon, '.png') !== false ) :
-
-				$re  = '<span class="png-icon ' . sanitize_title( $icon ) . '">';
-				$re .= '<img src="' . get_stylesheet_directory_uri() . '/dist/icons/'. $icon .'" alt="'. preg_replace('/\\.[^.\\s]{3,4}$/', '', $icon) .'" />';
-				$re .= '</span>';
-
-			elseif ( $original ) :
-
-				$src = get_stylesheet_directory_uri() . '/dist/icons/'. $icon .'.svg';
-				$replace = 'Layer' . rand();
-
-				$re  = '<span class="svg-icon ' . sanitize_title( $icon ) . '">';
-				$re .= str_replace('id="Layer_1"', $replace, file_get_contents($src));
-				$re .= '</span>';
-
-			else :
-
-				$re  = '<span class="svg-icon ' . sanitize_title( $icon ) . '">';
-				$re .= '<svg class="svg-'. $icon .'">';
-				$re .= '<use xlink:href="'.get_stylesheet_directory_uri().'/dist/sprite/spritemap.svg#sprite-'. $icon .'"></use>';
-				$re .= '</svg>';
-				$re .= '</span>';
-
-			endif;
-
-			if ( $return )
-				return $re;
-
-			echo $re;
+		if ( ! $icon ) {
+			$icon = get_key('select_icon');
 		}
-	// endif;
+
+		if ( is_array($icon) ) {
+			$icon = $icon['select_icon'];
+		}
+
+		if ( strpos($icon, '.png') !== false ) :
+
+			$re  = '<span class="png-icon ' . sanitize_title( $icon ) . '">';
+			$re .= '<img src="' . get_stylesheet_directory_uri() . '/dist/icons/'. $icon .'" alt="'. preg_replace('/\\.[^.\\s]{3,4}$/', '', $icon) .'" />';
+			$re .= '</span>';
+
+		elseif ( $original ) :
+
+			$src = get_stylesheet_directory_uri() . '/dist/icons/'. $icon .'.svg';
+			$replace = 'Layer' . rand();
+
+			$re  = '<span class="svg-icon ' . sanitize_title( $icon ) . '">';
+			$re .= str_replace('id="Layer_1"', $replace, file_get_contents($src));
+			$re .= '</span>';
+
+		else :
+
+			$re  = '<span class="svg-icon ' . sanitize_title( $icon ) . '">';
+			$re .= '<svg class="svg-'. $icon .'">';
+			$re .= '<use xlink:href="'.get_stylesheet_directory_uri().'/dist/sprite/spritemap.svg#sprite-'. $icon .'"></use>';
+			$re .= '</svg>';
+			$re .= '</span>';
+
+		endif;
+
+		if ( $return )
+			return $re;
+
+		echo $re;
+	}
 
 
 
@@ -339,24 +283,23 @@ class Helpers {
 	/**
 	 * Include ACF Youtube oembed
 	 */
-	// if(!function_exists('get_acf_oembed')) {
-		function get_acf_oembed( $field ) {
+	function get_acf_oembed( $field ) {
 
-			if( empty( $field ) ) return;
+		if( empty( $field ) ) return;
 
-			preg_match('%(?:youtube(?:-nocookie)?\.com/(?:[^/]+/.+/|(?:v|e(?:mbed)?)/|.*[?&]v=)|youtu\.be/)([^"&?/ ]{11})%i', $field, $match);
+		preg_match('%(?:youtube(?:-nocookie)?\.com/(?:[^/]+/.+/|(?:v|e(?:mbed)?)/|.*[?&]v=)|youtu\.be/)([^"&?/ ]{11})%i', $field, $match);
 
-			if ( empty($match[1]) ) return $field;
+		if ( empty($match[1]) ) return $field;
 
-			return sprintf(
-				'<iframe width="640" height="360"
-					src="https://www.youtube-nocookie.com/embed/%s?feature=oembed"
-					frameborder="0"
-					allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture"
-					allowfullscreen>
-				</iframe>',
-				$match[1]
-			);
-		}
-	// }
+		return sprintf(
+			'<iframe width="640" height="360"
+				src="https://www.youtube-nocookie.com/embed/%s?feature=oembed"
+				frameborder="0"
+				allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture"
+				allowfullscreen>
+			</iframe>',
+			$match[1]
+		);
+	}
+
 }
