+++ inc/utils/wrapper.php
@@ -2,7 +2,8 @@
 
 namespace CS\Utils;
 
-if ( ! defined( 'ABSPATH' ) ) exit;
+if ( ! defined( 'ABSPATH' ) )
+	exit;
