--- inc/wp-mods/wp-cleanup.php
@@ -1,71 +1,0 @@
-<?php if (!defined('ABSPATH')) exit;
-
-/**
- * Clean up wp-head
- */
-add_action('after_setup_theme', function () {
-
-    remove_action('wp_head', 'rsd_link');
-    remove_action('wp_head', 'wp_generator');
-    remove_action('wp_head', 'feed_links', 2);
-    remove_action('wp_head', 'feed_links_extra', 3);
-    remove_action('wp_head', 'index_rel_link');
-    remove_action('wp_head', 'wlwmanifest_link');
-    remove_action('wp_head', 'start_post_rel_link', 10, 0);
-    remove_action('wp_head', 'parent_post_rel_link', 10, 0);
-    remove_action('wp_head', 'adjacent_posts_rel_link', 10, 0);
-    remove_action('wp_head', 'adjacent_posts_rel_link_wp_head', 10, 0);
-    remove_action('wp_head', 'wp_shortlink_wp_head', 10, 0);
-    remove_action('wp_head', 'rel_canonical');
-    remove_action('wp_head', 'rel_alternate');
-    remove_action('wp_head', 'wp_oembed_add_discovery_links');
-    remove_action('wp_head', 'wp_oembed_add_host_js');
-    remove_action('wp_head', 'rest_output_link_wp_head');
-
-    remove_action('rest_api_init', 'wp_oembed_register_route');
-    remove_action('wp_print_styles', 'print_emoji_styles');
-
-    remove_filter('oembed_dataparse', 'wp_filter_oembed_result', 10);
-    remove_filter('pre_oembed_result', 'wp_filter_pre_oembed_result', 10);
-
-    add_filter('embed_oembed_discover', '__return_false');
-
-    if ( ! current_user_can('manage_options') && ! is_admin() )
-        show_admin_bar(false);
-});
-
-
-/**
- * Disable the emoji's
- */
-add_action( 'init', function () {
-
-    remove_action( 'wp_head', 'print_emoji_detection_script', 7 );
-    remove_action( 'admin_print_scripts', 'print_emoji_detection_script' );
-    remove_action( 'wp_print_styles', 'print_emoji_styles' );
-    remove_action( 'admin_print_styles', 'print_emoji_styles' );
-    remove_filter( 'the_content_feed', 'wp_staticize_emoji' );
-    remove_filter( 'comment_text_rss', 'wp_staticize_emoji' );
-    remove_filter( 'wp_mail', 'wp_staticize_emoji_for_email' );
-
-
-    add_filter( 'tiny_mce_plugins', function ( $plugins ) {
-
-        if ( is_array( $plugins ) ) {
-            return array_diff( $plugins, array( 'wpemoji' ) );
-        } else {
-            return array();
-        }
-    });
-
-
-    add_filter( 'wp_resource_hints', function ( $urls, $relation_type ) {
-
-        if ( 'dns-prefetch' == $relation_type ) {
-            $emoji_svg_url = apply_filters( 'emoji_svg_url', 'https://s.w.org/images/core/emoji/2/svg/' );
-            $urls          = array_diff( $urls, array( $emoji_svg_url ) );
-        }
-
-        return $urls;
-    }, 10, 2 );
-});
